+++ Snyk.VisualStudio.Extension.2022/Authentication/AuthenticationToken.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Text.RegularExpressions;
 
 namespace Snyk.VisualStudio.Extension.Authentication
 {
@@ -35,6 +36,8 @@ namespace Snyk.VisualStudio.Extension.Authentication
             {
                 case AuthenticationType.Token:
                     return Guid.TryParse(this.value, out _);
+                case AuthenticationType.Pat:
+                    return Regex.IsMatch(this.value, @"^snyk_(?:uat|sat)\.[a-z0-9]{8}\.[a-zA-Z0-9-_]+\.[a-zA-Z0-9-_]+$^snyk_(?:uat|sat)\.[a-z0-9]{8}\.[a-zA-Z0-9-_]+\.[a-zA-Z0-9-_]+$"); ;
                 case AuthenticationType.OAuth:
                     {
                         var tokenState = GetTokenState(this.value);