+++ Snyk.VisualStudio.Extension.2022/Authentication/AuthenticationType.cs
@@ -4,9 +4,11 @@ namespace Snyk.VisualStudio.Extension.Authentication
 {
     public enum AuthenticationType
     {
-        [Description("OAuth2 authentication")]
+        [Description("OAuth2 (Recommended)")]
         OAuth,
-        [Description("Token authentication")]
+        [Description("Personal Access Token")]
+        Pat,
+        [Description("API Token (Legacy)")]
         Token,
     }
 }