+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.3.0]
 ### Changed
 - PAT support.
+- remove Snyk Code Quality
 
 ## [2.2.1]
 ### Fixed